✨ feat: 支持 Webhook 私网目标投递并增强 star 同步能力
- 通知渠道启用 allow_private 时，请求携带私网放行标记，拨号阶段允许私网/回环地址但仍拦截 169.254.x.x 与未指定地址，降低 SSRF 风险
- star 枚举客户端透传 external_pat，配额由匿名 60 次/小时提升至 5000 次/小时，降低大批量 star 同步的限流概率
- 手动触发及配置变更后的 star 同步改为后台异步执行，避免多仓库分页拉取阻塞 HTTP 请求导致超时
- ResourceIdentity 的 number 参数统一为 int64，修复 32 位系统下 Release 等大 ID 被截断导致指纹冲突
- 修复 digest 预览行空格缺失与同分事件排序不稳定，并补充 Agent 发现的 /settings 与 /starred-releases 路径

+++ internal/normalizer/fingerprint.go
@@ -30,8 +30,8 @@ func StateHash(parts ...string) string {
 	return hex.EncodeToString(sum[:16])
 }
 
-// ResourceIdentity 生成资源标识字符串。
-func ResourceIdentity(kind string, number int, runID int64) string {
+// ResourceIdentity 生成资源标识字符串。number 采用 int64 避免 32 位系统截断大 ID（如 Release ID）。
+func ResourceIdentity(kind string, number int64, runID int64) string {
 	if runID != 0 {
 		return fmt.Sprintf("run:%d", runID)
 	}
